requirements.txt
accelerate==0.20.3
evaluate==0.4.0
fqdn==1.5.1
isoduration==20.11.0
jsonpointer==2.3
jupyter==1.0.0
langchain==0.0.271
nltk==3.8.1
openai==0.28.0
pip-chill==1.0.3
protobuf==3.20.0
python-dotenv==1.0.0
sentencepiece==0.1.99
seqeval==1.2.2
torchaudio==2.0.2
torchvision==0.15.2
transformers==4.30.2
uri-template==1.2.0
webcolors==1.13
click~=8.1.6
pydantic~=1.10.12
tqdm~=4.65.0
pyyaml~=6.0.1
torch~=2.0.1
scikit-learn~=1.3.0
setuptools~=68.0.0
faiss-cpu~=1.7.4
scipy~=1.11.1